Changes for CAPI cluster upgrade
Etcd machines get upgraded first so that the controlplane spec gets updated etcd
endpoints. But the outdated controlplane machines still have the etcd endpoints
of the older etcd members. So for the upgrade to proceed, the older etcd members
need to still be a part of the etcd cluster. As a solution in absence of static
etcd endpoints, during etcd upgrade, after adding a new etcd member, instead of
deleting an older member it will only be removed as an owned machine. This way
the existing oudated controlplane VMs can still connect to the etcd servers
they were started with. And once the controlplane upgrade is complete, the older
outdated etcd members will be deleted