Modify Initialized condition check
Currently controller checks the initialized condition and tries to set it to
true only if number of current machines is 1. This works fine during creation.
But after move as the etcd cluster is catching up with existing machines,
there's an intermittent failure caused by setting it to false when first machine is detected,
and then next reconcile immediately could detect 2 machines, so this particular
check is never evaluated and initialized never gets set to true.
Instead the condition will now check if the number of replicas is greater than 0
and set the Initialized condition to true.