Save etcd http client separately per cluster
Each etcdadmCluster object needs to use a separate http client, since
this client contains the etcd cluster's ca cert and client cert-key pair.
This commit replaces the field containing http client with a sync.Map
where key is the cluster UID and value is the etcd http client.
It uses golang's sync Map because as per docs:
* The Map type is optimized for when the entry for a given key is only ever
 written once but read many times
* Use of a sync.Map may significantly reduce lock contention compared to a
Go map paired with a separate Mutex or RWMutex.
https://pkg.go.dev/sync#Map