Add endpoints available condition
The Ready condition on EtcdadmCluster is set as an aggregate of all conditions
set on the object. This is the logic that all CAPI controllers follow for the
Ready condition of the CRDs they manage.
During cluster creation, Initialized is the last condition that gets set on the
etcdadmCluster object. It gets set to true when the first etcd member is active.
So the Ready condition also gets set to true during cluster creation once the
Initialized condition becomes true.
This commit adds another condition on the EtcdadmCluster object
called `EtcdEndpointsAvailable` which becomes true only when all endpoints
pass healthcheck. It gets set to false before etcd cluster provisioning starts,
so that Ready condition as an aggregate will hold the value of `false` till all
endpoints are ready.