Disown machine if owning more than desired replicas
Under normal circumstances, the etcd machines owned by the etcdadm
cluster should never be higher than the specified number of desired
replicas, they should be equal at most.

However, it's possible that due to stale client caches or even manual
updates (where a user re-adds the owner reference to an old etcd
machine), an etcdadm cluster might appear to own (during an upgrade
reconcile loop) more machines that the number of desired replicas. In
that case, regardless of the reason, we want to remove the owner
reference before creating new replicas. If not, the next reconciliation
loop will still detect an owned machine out of spec and will create a
new replica, again without removing ownership of the out of spec
machine. This causes a loop of new machines being created without a
limit.